requirements.txt
pyTelegramBotAPI
selenium
webdriver-managernicorn
pytz